% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovNames.R
\name{getCovNames}
\alias{getCovNames}
\title{This function is just to help name parts of the acov and the vectorized components}
\usage{
getCovNames(varList)
}
\arguments{
\item{varList}{vector of names of the variables (in order!)}
}
\value{
list of names for rows and names for columns
}
\description{
This function is just to help name parts of the acov and the vectorized components
}
\examples{
getCovNames(x)

}
