% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage1Estimates_YB2007.R
\name{main}
\alias{main}
\title{Decomposes sigmaW sigmaB and estimates acov matrices as well}
\usage{
main(myData, clusterVar, varNames)
}
\arguments{
\item{myData}{description tbd}

\item{clusterVar}{description tbd}

\item{varNames}{description tbd}
}
\value{
sigmaw, sigmab, two acov matrices
}
\description{
Decomposes sigmaW sigmaB and estimates acov matrices as well
}
\examples{
main(myData, clusterVar, varNames)

}
