% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUMLdecomp.R
\name{MUMLdecomp}
\alias{MUMLdecomp}
\title{Muthen Decomp}
\usage{
MUMLdecomp(myData, clusterVar)
}
\arguments{
\item{myData}{dataframe}

\item{clusterVar}{name of the clustering variable}
}
\description{
Takes a DF and clustering variables and returns a list including decomposed covariance matrices
}
\details{
Details to be written
}
