% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterEstimatesTable.R
\name{parameterEstimatesTable}
\alias{parameterEstimatesTable}
\title{Takes a fit obj and returns a dataframe with the parameter estiamtes}
\usage{
parameterEstimatesTable(fitObj)
}
\arguments{
\item{fitObj}{MSEMMIIVSEM fit obj}
}
\value{
df
}
\description{
Takes a fit obj and returns a dataframe with the parameter estiamtes
}
\examples{
parameterEstimatesTable(fit)

}
