% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseInstrumentSyntax.R
\name{parseInstrumentSyntax}
\alias{parseInstrumentSyntax}
\title{Parse user-supplied instrumental variables}
\usage{
parseInstrumentSyntax(d, instruments, miiv.check)
}
\arguments{
\item{d}{tbd}

\item{instruments}{tbd}

\item{miiv.check}{tbd}
}
\description{
Parse user-supplied instrumental variables
}
\examples{
parseInstrumentSyntax(d, instruments, miiv.check)

}
\keyword{internal}
