% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMsemSyntax.R
\name{parseMSEMSyntax}
\alias{parseMSEMSyntax}
\title{Returns 2SLS estimates}
\usage{
parseMSEMSyntax(model)
}
\arguments{
\item{model}{lavaan style syntax}
}
\value{
list of level 1 and level 2 models
}
\description{
Returns 2SLS estimates
}
\examples{
parseMSEMSyntax(twolevelmodels)

}
