% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revVech.R
\name{revVech}
\alias{revVech}
\title{This helper function was lifted from MIIVsem}
\usage{
revVech(x, nrow = NULL)
}
\arguments{
\item{x}{tbd}

\item{nrow}{tbd}
}
\value{
returns a matrix
}
\description{
This helper function was lifted from MIIVsem
}
\examples{
revVech

}
