% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarg_jc2018.R
\name{sarg_jc2018}
\alias{sarg_jc2018}
\title{Corrected Sargan's test from Jin and Cao (2018).}
\usage{
sarg_jc2018(eqn_obj, fullCov, vCov, acov, Svv, Svz, Svy)
}
\arguments{
\item{vechCov}{vectorized covariance matrix}
}
\value{
list of sarg est, df, and p
}
\description{
Corrected Sargan's test from Jin and Cao (2018).
}
\examples{
sarg_jc2018(vechCov, varNames, dv, si, instru)

}
