% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage1Estimates_lavaan.R
\name{stage1Estimates_lavaan}
\alias{stage1Estimates_lavaan}
\title{Obtains stage 1 estimates with lavaan}
\usage{
stage1Estimates_lavaan(data, obsVars, clusterVar)
}
\arguments{
\item{data}{dataframe}

\item{obsVars}{names of the observed variables}

\item{clusterVar}{name of the cluster variable}
}
\value{
covariance matrices and the related acov matrices
}
\description{
Obtains stage 1 estimates with lavaan
}
\examples{
stage1Estimates_lavaan(data, obsVars, clusterVar)

}
