% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaSeSarg.R
\name{thetaSeSarg}
\alias{thetaSeSarg}
\title{All est steps together}
\usage{
thetaSeSarg(eqn_obj, fullCov, fullAcov)
}
\arguments{
\item{cov}{vectorized covariance matrix}

\item{varNames}{names of all variables (in order) from vechCov}

\item{dv}{dependent var}

\item{si}{scaling indicator}

\item{instru}{instument set}

\item{acov}{associated asymptotic covariance matrix for cov}
}
\value{
estimates
}
\description{
All est steps together
}
\examples{
thetaSeSarg(vechCov, varNames, dv, si, instru)

}
