% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vech.R
\name{vech}
\alias{vech}
\title{Takes a square matrix and returns a vector of unique elements}
\usage{
vech(myMatrix)
}
\arguments{
\item{myMatrix}{A square matrix}
}
\value{
vector of unique elements
}
\description{
Takes a square matrix and returns a vector of unique elements
}
\examples{
vech(df, 2018, 08)

}
