% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSatLavModel.R
\name{writeSatLavModel}
\alias{writeSatLavModel}
\title{Takes a list of observed variables and makes writes the lavaan syntax for a
fully saturated lavaan model}
\usage{
writeSatLavModel(obsVars)
}
\arguments{
\item{obsVars}{a vector of names of the observed vars in a model}
}
\value{
lav model syntax for saturated two level model
}
\description{
Takes a list of observed variables and makes writes the lavaan syntax for a
fully saturated lavaan model
}
\examples{
writeSatLavModel(obsVars)

}
