% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yb2007_ascov.R
\name{yb2007_ascov}
\alias{yb2007_ascov}
\title{CalculatiNG the asymptotic covariance matrix of the saturated model;
UsiNG only the level2 units as buildiNG blocks;}
\usage{
yb2007_ascov(beta, p, dup, nlevel1, NG, smatw, ymean, vsmatL1, varNames)
}
\arguments{
\item{beta}{description tbd}

\item{p}{description tbd}

\item{dup}{description tbd}

\item{nlevel1}{description tbd}

\item{NG}{description tbd}

\item{smatw}{description tbd}

\item{ymean}{description tbd}

\item{vsmatL1}{description tbd}
}
\value{
list of amat and gamma
}
\description{
CalculatiNG the asymptotic covariance matrix of the saturated model;
UsiNG only the level2 units as buildiNG blocks;
}
\examples{
yb2007_ascov(df, 2018, 08)

}
