% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yb2007_dp.R
\name{yb2007_DP}
\alias{yb2007_DP}
\title{takes an argument 'p' and returns a duplication matrix as defined in Magnus&Neudecker;
also see matrixcalc::duplication.matrix() for a possibly faster implementation}
\usage{
yb2007_DP(p)
}
\arguments{
\item{p}{any scalar}
}
\value{
matrix. duplication matrix of 1's and 0's
}
\description{
takes an argument 'p' and returns a duplication matrix as defined in Magnus&Neudecker;
also see matrixcalc::duplication.matrix() for a possibly faster implementation
}
\examples{
yb2007_DP(6)

}
