% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yb2007_mdm1.R
\name{yb2007_mdm1}
\alias{yb2007_mdm1}
\title{helper function, which takes beta (vector of all estimates) and converts
mu vector, sigw matrix, and sigb matrix;}
\usage{
yb2007_mdm1(p, beta, varNames)
}
\arguments{
\item{p}{scalar value. # of variables}

\item{beta}{vector of all estimates in order from mu, sigw, sigb}
}
\value{
list
}
\description{
helper function, which takes beta (vector of all estimates) and converts
mu vector, sigw matrix, and sigb matrix;
}
\examples{
yb2007_mdm1(5, beta)

}
