% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yb2007_minq1.R
\name{yb2007_minq1}
\alias{yb2007_minq1}
\title{**Maximizing the log likelihood function for the 2-level saturated model;}
\usage{
yb2007_minq1(dup, nlevel1, NG, beta0, smatw, ymean, vsmatL1, varNames)
}
\arguments{
\item{dup}{description tbd}

\item{nlevel1}{description tbd}

\item{NG}{description tbd}

\item{beta0}{description tbd}

\item{smatw}{description tbd}

\item{ymean}{description tbd}

\item{vsmatL1}{description tbd}
}
\value{
list
}
\description{
**Maximizing the log likelihood function for the 2-level saturated model;
}
\examples{
yb2007_minq1(a, b, c, d, e, f, g)

}
