% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yb2007_permuteGamma.R
\name{yb2007_permuteGamma}
\alias{yb2007_permuteGamma}
\title{creates a permutation matrix so that the order of elements in the Gamma matrix are rearranged}
\usage{
yb2007_permuteGamma(p)
}
\arguments{
\item{p}{description tbd}
}
\value{
matrix
}
\description{
creates a permutation matrix so that the order of elements in the Gamma matrix are rearranged
}
\examples{
yb2007_permuteGamma(6)

}
