% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yb2007_processData.R
\name{yb2007_processData}
\alias{yb2007_processData}
\title{Subroutine that process the data}
\usage{
yb2007_processData(ymat, p, nlevel1, NG)
}
\arguments{
\item{ymat}{description tbd}

\item{p}{description tbd}

\item{nlevel1}{description tbd}

\item{NG}{description tbd}
}
\value{
list
}
\description{
Subroutine that process the data
}
\examples{
yb2007_processData()

}
